feat: Allow manual estimated wait time on queue creation
Adds an optional `estimatedWaitTime` parameter to the `createQueue` service function and the `QueueInfo` type. This allows users to manually set an estimated wait time when creating a new queue, which will be used to initialize the queue's metrics.

This also updates the UI components (`CustomerView`, `DisplayView`, `QueueList`) to handle and display this new information where relevant.